+++ tests/__init__.py
@@ -3,3 +3,4 @@ from . import test_tjp_export
 from . import test_tjp_schedule_import
 from . import test_scheduler
 from . import test_gantt
+from . import test_scenario_selection